+++ packages/core/test/ArraySnap.test.jsx
@@ -1,4 +1,4 @@
 import Form from '../src';
-import arrayTests from '../testSnap/arrayTests';
+import { arrayTests } from '@rjsf/snapshot-tests';
 
 arrayTests(Form);