+++ packages/fluentui-rc/src/index.ts
@@ -1,0 +1,9 @@
+import FluentForm from './FluentForm';
+
+export { default as Form, generateForm } from './FluentForm';
+export { default as Templates, generateTemplates } from './Templates';
+export { default as Theme, generateTheme } from './Theme';
+export { default as Widgets, generateWidgets } from './Widgets';
+export { __createFluentUIRCFrameProvider } from './FluentUIRCFrameProvider';
+
+export default FluentForm;